linux-dmabuf: Fix rendering issue introduced in commit 9e6529c
As the modifiers are collected in a uint64_t array, the function called for each
item of the array needs to use the same type so that bit shift operations don't
result to invalid modifiers.

Issue detected by the mighty Žan Doberšek.

+++ src/linux-dmabuf/linux-dmabuf.cpp
@@ -370,7 +370,7 @@ bind_linux_dmabuf(struct wl_client *client, void *data, uint32_t version, uint32
     wl_resource_set_implementation(resource, &linux_dmabuf_implementation,
                                    data, NULL);
 
-    WS::Instance::singleton().foreachDmaBufModifier([version, resource] (int format, int modifier) {
+    WS::Instance::singleton().foreachDmaBufModifier([version, resource] (int format, uint64_t modifier) {
         if (version >= ZWP_LINUX_DMABUF_V1_MODIFIER_SINCE_VERSION) {
             uint32_t modifier_lo = modifier & 0xFFFFFFFF;
             uint32_t modifier_hi = modifier >> 32;
